+++ package.json
@@ -10,6 +10,7 @@
     "react-dom": "^17.0.2",
     "react-hooks": "^1.0.1",
     "react-redux": "^7.2.4",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
     "styled-components": "^5.3.0",
