+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import AppWrapper from './App';
+import App from './App';
 import styled from 'styled-components';
 
 const Container = styled.div`
@@ -15,7 +15,7 @@ const Container = styled.div`
 
 ReactDOM.render(
   <Container>
-    <AppWrapper/>
+    <App/>
   </Container>,
   document.getElementById('root')
 );
